refactor: 컴포넌트 폴더 깊이 단순화
