+++ src/components/pages/Main/CategoriesSection.css
