+++ src/components/pages/Main/index.js
@@ -1,1 +1,3 @@
-export { default as MainPage } from './MainPage';
+export { default as CategoriesSection } from './CategoriesSection';
+export { default as ContactSection } from './ContactSection';
+export { default as InfoSection } from './InfoSection';
