+++ src/pages/MainPage.jsx
@@ -1,10 +1,8 @@
 import React, { useState, useEffect } from 'react';
-import Header from '../components/layout/Header/Header';
-import HeroSection from '../components/layout/HeroSection/HeroSection';
-import CategoriesSection from '../components/pages/Main/CategoriesSection/CategoriesSection';
-import InfoSection from '../components/pages/Main/InfoSection/InfoSection';
-import ContactSection from '../components/pages/Main/ContactSection/ContactSection';
-import Footer from '../components/layout/Footer/Footer';
+import { Header } from '../components/layout';
+import { HeroSection } from '../components/layout';
+import { CategoriesSection, InfoSection, ContactSection } from '../components/pages/Main';
+import { Footer } from '../components/layout';
 import { mailAPI, googleFormsAPI } from '../services/api/index.js';
 import './MainPage.css';
